+++ walkers/src/map.rs
@@ -178,6 +178,7 @@ impl<'a, 'b, 'c, P: Projection + 'static> Map<'a, 'b, 'c, P> {
         for layer in self.layers {
             draw_tiles(
                 &painter,
+                &self.projection,
                 map_center,
                 zoom,
                 layer.tiles,